+++ src/configs/index.ts
@@ -3,11 +3,6 @@ import {
   EEscrowAddress,
   EChains
 } from '../types'
-export const polygonApiUrl = 'https://escrow-api.linkdrop.io/v2/polygon'
-export const mumbaiApiUrl = 'https://escrow-api.linkdrop.io/v2/mumbai'
-export const baseApiUrl = 'https://escrow-api.linkdrop.io/v2/base'
-export const baseGoeliApiUrl = 'https://escrow-api.linkdrop.io/v2/base-goerli'
-export const devApiUrl = 'https://escrow-api.linkdrop.io/dev'
 
 export const polygonJSONRPCUrl = 'https://polygon-mainnet.infura.io/v3/f5f8608d75874e5fbcdd4ee9657b8ab5'
 export const mumbaiJSONRPCUrl = 'https://polygon-mumbai.infura.io/v3/f5f8608d75874e5fbcdd4ee9657b8ab5'
@@ -15,6 +10,7 @@ export const baseJSONRPCUrl = 'https://developer-access-mainnet.base.org'
 export const baseGoerliJSONRPCUrl = 'https://base-goerli.publicnode.com'
 
 export const baseUrl = 'http://localhost:3000'
+export const apiUrl = 'https://escrow-api.linkdrop.io/v2'
 
 export const nativeTokenAddress = '0x0000000000000000000000000000000000000000'