+++ src/modules/linkdrop-p2p/types/constructor-args/index.ts
@@ -1,6 +1,7 @@
 type TConstructorArgs = {
-  apiKey: string
+  apiKey?: string
   baseUrl: string
+  apiUrl?: string
 }
 
 export default TConstructorArgs